initial test programs for new canvas controller
