+++ templates/analyze.html
@@ -2,6 +2,7 @@
 {% extends 'base.html' %}
 
 {% block head %}
+<script type="text/javascript" charset="utf8" src="static/canvas_controller.js"></script>
 <script type="text/javascript" charset="utf8" src="static/analyze.js"></script>
 <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/purecss@3.0.0/build/grids-responsive-min.css" />
 {% endblock %}
